requirements.txt
Django
python-dotenv==1.0.0
psycopg
psycopg2
djangorestframework
django-filter

channels
channels_redis
django-cors-headers
dj-rest-auth[with_social]
redis
djangorestframework-simplejwt
drf-yasg